layout/generic/test/test_bug405178.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=405178
-->
<head>
  <title>Test for Bug 405178</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=405178">Mozilla Bug 405178</a>
<p id="display"><div id="div" style="-moz-column-width: 10px;">&#23377;&#1741; </div></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 405178 **/

var div, textNode;

function boom()
{
  div = document.getElementById("div");
  textNode = div.firstChild;

  div.removeChild(textNode);
  setTimeout(boom2, 200);
}

function boom2()
{
  textNode.data = "";
  div.appendChild(document.createTextNode("X"))
  var foo = document.body.offsetHeight;

  ok(true, "Test is successful if we get here without crashing");
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
SimpleTest.requestFlakyTimeout("untriaged");
addLoadEvent(boom);

</script>
</pre>
</body>
</html>
